+++ unipressed/id_mapping/core.py
@@ -4,10 +4,18 @@ from dataclasses import dataclass
 from typing import Iterable
 
 import requests
-from typing_extensions import TypeAlias, TypedDict
+from typing_extensions import Literal, TypeAlias, TypedDict
 
 from unipressed.id_mapping.types import From, To
 from unipressed.util import iter_pages
+
+UniprotStatus: TypeAlias = Literal[
+    # See https://github.com/ebi-uniprot/uniprot-rest-api/blob/master/idmapping-rest/src/main/java/org/uniprot/api/idmapping/controller/response/JobStatus.java
+    "NEW",
+    "RUNNING",
+    "FINISHED",
+    "ERROR",
+]
 
 
 class IdMappingError(Exception):
@@ -45,13 +53,23 @@ class IdMappingJob:
 
     job_id: str
 
-    def get_status(self) -> str:
+    def get_status(self) -> UniprotStatus:
         """
         Returns a string describing the status of the job (ie if it has completed or not).
         """
-        return requests.get(
+        data = requests.get(
             f"https://rest.uniprot.org/idmapping/status/{self.job_id}"
-        ).json()["jobStatus"]
+        ).json()
+        if "jobStatus" in data:
+            # This is the correct behaviour
+            return data["jobStatus"]
+        elif "results" in data:
+            # Sometimes UniProt returns the actual results even though we asked for the status, see https://github.com/multimeric/Unipressed/issues/21
+            return "FINISHED"
+        else:
+            raise IdMappingError(
+                f"Unknown response returned by UniProt. Received {data}"
+            )
 
     def each_result(self) -> Iterable[IdMappingResult]:
         """